requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
timm
ultralytics
requests
python-multipart
